// src/fsd/3-features/guides/guides.enums.ts
export enum GuidesStatus {
    pending = 0,
    approved = 1,
    rejected = 2,
}

export enum SlotType {
    core = 0,
    flex = 1,
    none = 2,
}

export enum GuidesGroup {
    approved = 0,
    favorite = 1,
    myTeams = 2,
    pending = 3,
    rejected = 4,
}

export enum LreCharacter {
    kharn = '_kharn',
    vitruvius = '_vitruvius',
    mephiston = '_mephiston',
    patermine = '_patermine',
    dante = '_dante',
}